Remove failback. Update logger. Add loop to drain messages in TransferMessages function

+++ ServiceBusDR/Activities/InitiatePairingActivity.cs
@@ -3,6 +3,8 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Microsoft.Extensions.Logging;
 using ServiceBusDR.Services;
+using System;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace ServiceBusDR
@@ -10,16 +12,14 @@ namespace ServiceBusDR
     public class InitiatePairingActivity
     {
         private readonly IGeoService _geoService;
-        private readonly ILogger _logger;
 
-        public InitiatePairingActivity(IGeoService geoService, ILogger<InitiatePairingActivity> logger)
+        public InitiatePairingActivity(IGeoService geoService)
         {
             _geoService = geoService;
-            _logger = logger;
         }
 
         [FunctionName(nameof(InitiatePairingActivity))]
-        public async Task Run([ActivityTrigger] object input)
+        public async Task Run([ActivityTrigger] object input, ILogger logger)
         {
             var geo = await _geoService.GetGeoNamespace();
 
@@ -27,35 +27,33 @@ namespace ServiceBusDR
             var isAlreadyPaired = pairingStatus.Role != RoleDisasterRecovery.PrimaryNotReplicating;
             if (isAlreadyPaired)
             {
-                _logger.LogInformation("Already paired. No action performed.");
+                logger.LogInformation("Already paired. No action performed.");
                 return;
             }
 
-            _logger.LogInformation("Attempting to create pairing");
+            logger.LogInformation("Attempting to create pairing");
 
-            await _geoService.TransferMessages(geo);
-
-            var isEmpty = await _geoService.DeleteAllEntities(geo);
-            while (!isEmpty)
+            var hasMessages = (await _geoService.GetNonEmptyEntities(geo)).Any();
+            if (hasMessages)
             {
-                _logger.LogInformation($"Cannot delete entities of '{geo.Partner.Name}' because some subscriptions contain messages");
-                _logger.LogInformation($"Transfering messages");
-                await _geoService.TransferMessages(geo);
-                isEmpty = await _geoService.DeleteAllEntities(geo);
+                var errorMessage = $"Cannot delete entities of '{geo.Partner.Name}' because some subscriptions contain messages. You have to run TransferMessages first.";
+                logger.LogError(errorMessage);  
+                throw new Exception(errorMessage);
             }
 
-            _logger.LogInformation("Starting Pairing");
+            await _geoService.DeleteAllEntities(geo);
+           
+            logger.LogInformation("Starting Pairing");
             var drStatus = await _geoService.InitiatePairing(geo);
 
-            _logger.LogInformation($"Waiting for pairing to complete");
+            logger.LogInformation($"Waiting for pairing to complete");
             var status = drStatus.ProvisioningState;
             while (status != ProvisioningStateDR.Succeeded)
             {
                 await Task.Delay(15000);
                 status = (await _geoService.GetPairingStatus(geo.Current.ResourceGroup, geo.Current.Name)).ProvisioningState;
             }
-
-            _logger.LogInformation("Pairing successful");
+            logger.LogInformation("Pairing successful");
         }
     }
 }
